requirements.txt
# Rasa and NLP
rasa==3.6.16
rasa-sdk==3.6.2

# Claude API
anthropic==0.34.2

# Backend
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
python-multipart==0.0.12

# Database
asyncpg==0.30.0
redis==5.2.0
sqlalchemy==2.0.36
alembic==1.14.0

# Integrations
hubspot-api-client==9.2.2
google-api-python-client==2.155.0
google-auth==2.36.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Utilities
python-dateutil==2.9.0.post0
pytz==2024.2
python-dotenv==1.0.1
httpx==0.28.1
tenacity==9.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# Development
black==24.10.0
flake8==7.1.1
isort==5.13.2
mypy==1.13.0